Refactor and split sm-api into two modules
* datasets.py and databases.py
